backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
python-multipart==0.0.12
python-dotenv==1.0.1

# Gemini + Google Cloud
google-genai==1.0.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2

# Document parsing
pymupdf==1.24.11
python-pptx==1.0.2
python-docx==1.1.2
